Move layout.js
