+++ gatsby-config.js
@@ -1,6 +1,6 @@
 module.exports = {
   siteMetadata: {
-    title: 'Gatsby Default Starter',
+    title: 'pinvite',
   },
   plugins: [
     'gatsby-plugin-react-helmet',
