+++ src/pages/index.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import {ButtonTwitter} from '../components/Atoms/Button'
 import ApplicationBar from '../components/Molecules/ApplicationBar'
-import Layout from '../components/layout'
+import Layout from '../components/Templates/Layout'
 import InviteSamplePaper from '../components/Molecules/InviteSamplePaper'
 import CallToAction from '../components/Molecules/CallToAction'
 import IssueListPaper from '../components/Molecules/IssueListPaper'
