+++ src/pages/invite-preview.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Link } from 'gatsby'
 
-import Layout from '../components/layout'
+import Layout from '../components/Templates/Layout'
 import cloudinary from '../utils/cloudinary'
 
 const InvitePreviewPage = () => {
