+++ src/pages/promote.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Link } from 'gatsby'
 
-import Layout from '../components/layout'
+import Layout from '../components/Templates/Layout'
 
 const PromotePage = () => (
   <Layout>
